release of methods for Monster_AI

+++ SimpleRPG/NumGen.cpp
@@ -2,6 +2,6 @@
 #include <ctime>
 int NumGen(int min, int max)
 {
-	srand(static_cast<unsigned long long>(time(0)));
+	srand(static_cast<unsigned int>(time(0)));
 	return (rand() % (max - min + 1) + min);
 }
